SlaySpire/requirements.txt
opencv-python
pyautogui
numpy
Pillow
pytesseract